Move Box components into directories
